+++ src/main/java/com/mraof/minestuck/alchemy/recipe/generator/recipe/RecipeGeneratedCostHandler.java
@@ -7,11 +7,11 @@ import com.google.gson.stream.JsonReader;
 import com.mraof.minestuck.Minestuck;
 import com.mraof.minestuck.api.alchemy.GristSet;
 import com.mraof.minestuck.api.alchemy.ImmutableGristSet;
-import com.mraof.minestuck.alchemy.recipe.generator.GeneratedCostProvider;
+import com.mraof.minestuck.api.alchemy.recipe.generator.GeneratedCostProvider;
 import com.mraof.minestuck.alchemy.recipe.generator.GenerationContext;
-import com.mraof.minestuck.alchemy.recipe.generator.GristCostResult;
-import com.mraof.minestuck.alchemy.recipe.generator.LookupTracker;
-import com.mraof.minestuck.jei.JeiGristCost;
+import com.mraof.minestuck.api.alchemy.recipe.generator.GristCostResult;
+import com.mraof.minestuck.api.alchemy.recipe.generator.LookupTracker;
+import com.mraof.minestuck.api.alchemy.recipe.JeiGristCost;
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.resources.ResourceLocation;
