+++ src/main/java/com/mraof/minestuck/alchemy/recipe/generator/recipe/RecipeGeneratedCostProcess.java
@@ -5,8 +5,8 @@ import com.mraof.minestuck.MinestuckConfig;
 import com.mraof.minestuck.api.alchemy.GristSet;
 import com.mraof.minestuck.api.alchemy.ImmutableGristSet;
 import com.mraof.minestuck.alchemy.recipe.generator.GenerationContext;
-import com.mraof.minestuck.alchemy.recipe.generator.GristCostResult;
-import com.mraof.minestuck.alchemy.recipe.generator.LookupTracker;
+import com.mraof.minestuck.api.alchemy.recipe.generator.GristCostResult;
+import com.mraof.minestuck.api.alchemy.recipe.generator.LookupTracker;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.crafting.Recipe;
 import org.apache.commons.lang3.tuple.Pair;
