+++ src/main/java/com/mraof/minestuck/api/alchemy/recipe/generator/GristCostResult.java
@@ -1,4 +1,4 @@
-package com.mraof.minestuck.alchemy.recipe.generator;
+package com.mraof.minestuck.api.alchemy.recipe.generator;
 
 import com.mraof.minestuck.api.alchemy.GristSet;
