+++ src/main/java/com/mraof/minestuck/jei/GristCostRecipeCategory.java
@@ -3,6 +3,7 @@ package com.mraof.minestuck.jei;
 import com.mraof.minestuck.alchemy.AlchemyHelper;
 import com.mraof.minestuck.api.alchemy.GristSet;
 import com.mraof.minestuck.api.alchemy.GristTypes;
+import com.mraof.minestuck.api.alchemy.recipe.JeiGristCost;
 import com.mraof.minestuck.block.MSBlocks;
 import com.mraof.minestuck.client.util.GuiUtil;
 import com.mraof.minestuck.player.ClientPlayerData;
