+++ src/main/java/com/mraof/minestuck/network/EditmodeDragPacket.java
@@ -1,10 +1,9 @@
 package com.mraof.minestuck.network;
 
 import com.mraof.minestuck.MinestuckConfig;
-import com.mraof.minestuck.alchemy.recipe.GristCostRecipe;
+import com.mraof.minestuck.api.alchemy.recipe.GristCostRecipe;
 import com.mraof.minestuck.api.alchemy.GristSet;
 import com.mraof.minestuck.api.alchemy.MutableGristSet;
-import com.mraof.minestuck.alchemy.recipe.GristCost;
 import com.mraof.minestuck.api.alchemy.GristTypes;
 import com.mraof.minestuck.computer.editmode.*;
 import com.mraof.minestuck.player.GristCache;
